Extract Observer surface to worker/log/observer.go (PR #32 review #7)
Per reviewer feedback: observer_test.go had no paired source file. Move the Observer type, observers atomic.Pointer, and RegisterObserver from handlers.go into a new observer.go so the file pairing is structurally explicit. Removes the now-unused sync/atomic import from handlers.go. No behavior change; all tests pass.
